use float instead int for the impacts, fix the initializer of the config db

+++ lib/ranking/reports.py
@@ -163,7 +163,7 @@ class Reports(CommonReport):
             if asn != self.config.get('modules_global','default_asn'):
                 rank = self.get_daily_rank(asn, source, date)
                 if rank is not None:
-                    pipeline.zincrby(histo_key, asn, float(rank) * int(self.config_db.get(str(source))))
+                    pipeline.zincrby(histo_key, asn, float(rank) * float(self.config_db.get(str(source))))
         pipeline.execute()
 
     def format_report(self, source = None, limit = 50, date = None):
@@ -246,9 +246,9 @@ class Reports(CommonReport):
                 if rank is not None:
                     asn_day = dates[i]
                     if ranks_by_days.get(asn_day, None) is None:
-                        ranks_by_days[asn_day] = float(rank) *  int(self.config_db.get(str(source)))
+                        ranks_by_days[asn_day] = float(rank) *  float(self.config_db.get(str(source)))
                     else:
-                        ranks_by_days[asn_day] += float(rank) *  int(self.config_db.get(str(source)))
+                        ranks_by_days[asn_day] += float(rank) *  float(self.config_db.get(str(source)))
                 i += 1 
         for ranks in ranks_by_days:
             ranks_by_days[ranks] += 1 
@@ -295,7 +295,7 @@ class Reports(CommonReport):
                 local_rank = 0.0
                 i = 0 
                 for source in sources:
-                    local_rank += float(ips_by_sources[i]) *  int(self.config_db.get(str(source)))
+                    local_rank += float(ips_by_sources[i]) *  float(self.config_db.get(str(source)))
                     i += 1
                 sources_web = self.history_db_temp.smembers(asn_timestamp_temp)
                 asn_descs_to_print.append([asn, timestamp, owner, ip_block,\
